requirements.txt
aiogram==3.4.1
gspread==5.11.2
oauth2client==4.1.3
python-dotenv==1.0.0